pyproject.toml
[tool.poetry]
name = "fastapi-mailing-service"
version = "0.1.0"
description = ""
authors = ["valhallajazzy <valhallajazzy@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pydantic-extra-types = "^2.4.1"
python-dotenv = "^1.0.1"
phonenumbers = "^8.13.29"
pytz = "^2023.4"
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
